Create README.md
The objective of BitCheques is to simplify the transfer and redemption of Bitcoin, eliminating the need to interact with the internet for day to day transactions while preventing unauthorized access to a private key, or requiring memorization of passwords, pass phrases, &c.

A person who receives a BitCheque in payment can take it to anyone with a btc wallet to exchange it for something else, or to a btc-enabled bank for deposit or fiat.

A BitCheque can be written in any amount of satoshis and loaded by scanning the QR code - similar to a paper wallet.

Instead of a private key on the cheque, a script hash is pre-generated and visible on the BitCheque as a separate QR code.

The script allows the funds to be transferred to an address or contract upon the verified signature of the holder. The issuer may not access the funds once they are loaded.

An signing party authorized by BitCheques (2of2) verifies the legitimacy of the holder and signs, releasing the funds.

Alternatively, an escrowed BitCheque can be set up (2of3) whereby an arbitrator chosen unanimously by the 2 parties - or another predetermined selection format - is authorized to sign.